refactor. handleConflicts in composable
